Fix animation hash self-references, attack index type, and Tank initialization
MechSuitAnimationController was assigning LegsRotation and LegsDead to
themselves instead of hashing the strings. EnemyAnimationController used
SetInteger for AttackIndex but the parameter is a float. Tank.OnValidate
moved to Start with a null check so it runs correctly at runtime.

+++ Assets/Scripts/Gameplay/Entity/Enemy/EnemyAnimationController.cs
@@ -117,7 +117,7 @@ public class EnemyAnimationController : EntityAnimationController
     public void OnAttack()
     {
         if (m_animator.GetBool(Dead)) return;
-        if (m_hasAttackIndex) m_animator.SetInteger(AttackIndex, CurrentAttackIndex);
+        if (m_hasAttackIndex) m_animator.SetFloat(AttackIndex, CurrentAttackIndex);
         m_animator.SetTrigger(Attack);
     }
